+++ data/colors/index.ts
@@ -1,17 +1,21 @@
 import light from './themes/light'
 import light_high_contrast from './themes/light_high_contrast'
 import light_colorblind from './themes/light_colorblind'
+import light_tritanopia from './themes/light_tritanopia'
 import dark from './themes/dark'
 import dark_dimmed from './themes/dark_dimmed'
 import dark_high_contrast from './themes/dark_high_contrast'
 import dark_colorblind from './themes/dark_colorblind'
+import dark_tritanopia from './themes/dark_tritanopia'
 
 export default {
   light,
   light_high_contrast,
   light_colorblind,
+  light_tritanopia,
   dark,
   dark_dimmed,
   dark_high_contrast,
-  dark_colorblind
+  dark_colorblind,
+  dark_tritanopia
 }