+++ data/colors/vars/component_dark.ts
@@ -167,5 +167,24 @@ export default {
       activeBg: alpha(get('scale.red.4'), 0.24),
       hoverText: get('scale.red.3')
     }
+  },
+
+  switchTrack: {
+    bg: get('scale.black'),
+    border: get('scale.gray.4'),
+
+    checked: {
+      bg: alpha(get('scale.blue.5'), 0.35),
+      hoverBg: alpha(get('scale.blue.5'), 0.5),
+      activeBg: alpha(get('scale.blue.5'), 0.65),
+      border: get('scale.blue.3'),
+    }
+  },
+
+  switchKnob: {
+    checked: {
+      bg: get('scale.blue.5'),
+      disabledBg: get('scale.gray.5'),
+    }
   }
 }